third_party/blink/web_tests/external/wpt/css/css-position/sticky/position-sticky-table-cell-scrollIntoView.html

<!DOCTYPE html>
<title>scrollIntoView on a sticky table cell uses its natural position</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#stickypos-scroll" />
<meta name="assert" content="scrollIntoView on a sticky-positioned table cell should scroll to the cell's natural layout position, not its stuck visual position" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
body { margin: 0; }
table { border-collapse: collapse; width: 100%; }
td:first-child {
  position: sticky;
  left: 0;
  top: 0;
}
td { height: 40px; }
</style>

<body>
<table>
  <tbody id="tbody"></tbody>
</table>
<script>
const tbody = document.getElementById("tbody");
const ROWS = 200;
const TARGET_ROW = 50;
let targetCell;
for (let i = 0; i < ROWS; i++) {
  const tr = document.createElement("tr");
  const td = document.createElement("td");
  td.textContent = i === TARGET_ROW ? "target" : i;
  tr.appendChild(td);
  tr.appendChild(document.createElement("td"));
  tbody.appendChild(tr);
  if (i === TARGET_ROW) targetCell = td;
}

test(() => {
  // Measure the cell's natural y at scrollY=0 before any sticky offset applies.
  window.scrollTo(0, 0);
  const naturalTop = targetCell.getBoundingClientRect().top + window.scrollY;

  // Scroll to the bottom so the cell is stuck at the top of the viewport.
  window.scrollTo(0, document.documentElement.scrollHeight);
  assert_greater_than(window.scrollY, naturalTop,
    "precondition: page is scrolled past the target cell");

  // scrollIntoView should scroll back to the cell's natural position.
  targetCell.scrollIntoView();
  assert_approx_equals(window.scrollY, naturalTop, 1,
    "scrollIntoView should scroll to the cell's natural layout position");
}, "scrollIntoView on a sticky table cell uses its natural layout position");
</script>
</body>
